Move events to utils
