+++ src/utils/events.js
@@ -1,4 +1,4 @@
-// Copyright Joyent, Inc. and other Node contributors.
+// Copyright Joyent, Inc. and other Node contributors. https://github.com/browserify/events
 //
 // Permission is hereby granted, free of charge, to any person obtaining a
 // copy of this software and associated documentation files (the
